requirements.txt
# General libraries
warnings
time
random

# OpenCV and Numpy for image and video processing
opencv-python-headless
numpy

# PyTorch for deep learning models
torch

# Librosa for audio and speech processing
librosa

# Requests and JSON for handling Ollama API interactions
requests
json

# HuggingFace Transformers for Wav2Vec2 models
transformers

# Pillow for image processing
Pillow

# Tkinter for file dialogs and GUI
tk